Refactoring and adding eslint
