+++ index.js
@@ -1,9 +1,9 @@
-var addSplitPoints = require("./split.js").addSplitPointsAll;
-var cornerPointMerger = require("./cornerPointMerger.js").cornerPointMerger;
-var generateCornerPoints = require("./split.js").generateCornerPoints;
-var generatePointSubset = require("./split.js").generatePointSubset;
-var mergeTiles = require("./mergeTiles.js").mergeTiles;
-var split = require("./split.js").split;
+var addSplitPoints = require("./src/split").addSplitPointsAll;
+var cornerPointMerger = require("./src/cornerPointMerger").cornerPointMerger;
+var generateCornerPoints = require("./src/split").generateCornerPoints;
+var generatePointSubset = require("./src/split").generatePointSubset;
+var mergeTiles = require("./src/mergeTiles").mergeTiles;
+var split = require("./src/split").split;
 
 module.exports = {
   addSplitPoints,
