+++ src/consoleManager.js
