+++ src/constants/gridPointTypes.js
