+++ src/cornerPointMerger.js
