+++ src/generateCornerPoints.js
