+++ src/inputAnalysis.js
