+++ src/mergeTiles.js
