+++ src/mergeTilesUtils/getCoordBoundary.js
