+++ src/mergeTilesUtils/getInnerPoints.js
