+++ src/split.js
